sensora service/setpointapi/requirements.txt
flask==3.0.2
solace-pubsubplus==1.5.0
